스텝 생성

+++ src/main/java/com/example/tily/roadmap/RoadmapRepository.java
@@ -2,6 +2,6 @@ package com.example.tily.roadmap;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoadmapRepository extends JpaRepository<Roadmap, Integer> {
+public interface RoadmapRepository extends JpaRepository<Roadmap, Long> {
 
 }
